Add support for relative and absolute links
