+++ components/Markdown.tsx
@@ -6,26 +6,14 @@ import marked, { Renderer } from "marked";
 import dompurify from "dompurify";
 import { RawCodeBlock } from "./CodeBlock";
 import { replaceEmojis } from "../util/emoji_util";
-import { markup, MarkupProps, scrollEffect, slugify } from "./Markup";
-
-function isRelative(path: string): boolean {
-  return (
-    !path.startsWith("/") &&
-    !path.startsWith("https://") &&
-    !path.startsWith("http://") &&
-    !path.startsWith("//")
-  );
-}
-
-function relativeToAbsolute(base: string, relative: string): string {
-  const baseURL = new URL(base);
-  baseURL.search = "";
-  baseURL.hash = "";
-  const parts = baseURL.pathname.split("/");
-  parts[parts.length - 1] = relative;
-  baseURL.pathname = parts.join("/");
-  return baseURL.href;
-}
+import {
+  markup,
+  MarkupProps,
+  scrollEffect,
+  slugify,
+  transformImageUri,
+  transformLinkUri,
+} from "./Markup";
 
 function Markdown(props: MarkupProps): React.ReactElement | null {
   scrollEffect();
@@ -110,45 +98,6 @@ function Markdown(props: MarkupProps): React.ReactElement | null {
     console.log(err);
     return null;
   }
-}
-
-function transformLinkUri(displayURL: string, baseURL: string) {
-  return (uri: string) => {
-    let href = uri;
-
-    if (uri.startsWith("#")) return uri;
-
-    // If the URL is relative, it should be relative to the canonical URL of the file.
-    if (isRelative(href)) {
-      // https://github.com/denoland/deno_website2/issues/1047
-      href = decodeURIComponent(relativeToAbsolute(displayURL, href));
-    }
-    if (href.startsWith("/") && !href.startsWith("//")) {
-      href = `${baseURL}${href}`;
-    }
-
-    const hrefURL = new URL(href);
-
-    // Manual links should not have trailing .md
-    if (
-      hrefURL?.pathname?.startsWith("/manual") &&
-      hrefURL?.origin === "https://deno.land"
-    ) {
-      hrefURL.pathname = hrefURL.pathname.replace(/\.md$/, "");
-      href = hrefURL.href;
-    }
-
-    return href;
-  };
-}
-
-function transformImageUri(sourceURL: string) {
-  return (uri: string) => {
-    if (isRelative(uri)) {
-      return relativeToAbsolute(sourceURL, uri);
-    }
-    return uri;
-  };
 }
 
 export default Markdown;
